+++ example/lib/main.dart
@@ -1,15 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_gemma/flutter_gemma.dart';
 import 'package:flutter_gemma_example/chat_screen.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  FlutterGemmaPlugin.instance.init(
-    maxTokens: 512,
-    temperature: 1.0,
-    topK: 1,
-    randomSeed: 1,
-  );
   runApp(const ChatApp());
 }